requirements.txt
# Core API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# AI & ML
openai==1.3.7
anthropic==0.7.8

# Graph Database
neo4j==5.15.0

# Caching & Queue
redis==5.0.1
celery==5.3.4

# Data Processing
pandas==2.1.4
numpy==1.26.2
python-dateutil==2.8.2
openpyxl==3.1.2

# Email & Communication
aiosmtplib==3.0.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
